requirements.txt
tensorflow
Pillow
numpy
scikit-learn
matplotlib